add support for sprites with no data file
